added formula classes

+++ Data/MeasurementHeader.cs
@@ -1,15 +1,15 @@
 ﻿namespace DICOMReporting.Data {
 	public struct MeasurementHeader {
 		public string Name;
-		public decimal Value;
+		public double Value;
 		public MeasurementUnit Unit;
 
-		public MeasurementHeader(string name, decimal value, MeasurementUnit unit = new MeasurementUnit()) {
+		public MeasurementHeader(string name, double value, MeasurementUnit unit = new MeasurementUnit()) {
 			Name = name;
 			Value = value;
 			Unit = unit;
 		}
-		public MeasurementHeader(string name, decimal value, string unitName, string unitShorthand) : this(name, value, new MeasurementUnit(unitName, unitShorthand)) {
+		public MeasurementHeader(string name, double value, string unitName, string unitShorthand) : this(name, value, new MeasurementUnit(unitName, unitShorthand)) {
 		}
 
 		public override string ToString() {
